refactor attributes
