--- Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/EditorOnly.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8cab898ac8c44ee6bc375690829e274d
-timeCreated: 1677336514
