--- Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/SerializableGlobalObjectId.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1d4f7ee187fd4529bcb879c64ccda957
-timeCreated: 1679921851
