--- Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/SpecifyObjectAttribute.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 031e3d77a4ff4f4b8652832e752101a5
-timeCreated: 1679922561
