+++ Kanikama/Packages/net.shivaduke28.kanikama.core/Runtime/Attributes/EditorOnlyAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Kanikama.Core
+namespace Kanikama.Core.Attributes
 {
     public sealed class EditorOnlyAttribute : Attribute
     {
