+++ Kanikama/Packages/net.shivaduke28.kanikama.gi.bakery/Baking/Editor/BakeryBakingSettingAsset.cs
@@ -1,5 +1,5 @@
 ﻿using System.IO;
-using Kanikama.Core.Editor;
+using Kanikama.Core.Editor.Util;
 using UnityEditor;
 using UnityEngine;
