+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Baking/Impl/KanikamaMonitor.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Kanikama.Core;
+using Kanikama.Core.Attributes;
 using UnityEngine;
 
 namespace Kanikama.GI.Baking.Impl
